pyproject.toml
[project]
name = "week-3-indiv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "geopandas>=1.1.1",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "shapely>=2.1.1",
    "uvicorn>=0.35.0",
]
